refactor!: rename autocli package
