+++ client/v2/autocli/cmd.go
@@ -1,4 +1,4 @@
-package cli
+package autocli
 
 import (
 	"errors"
