+++ client/v2/autocli/query.go
@@ -1,4 +1,4 @@
-package cli
+package autocli
 
 import (
 	"fmt"
