+++ client/v2/autocli/query_test.go
@@ -1,4 +1,4 @@
-package cli
+package autocli
 
 import (
 	"bytes"
