+++ client/v2/autocli/testdata/bank_example.yaml
