+++ client/v2/autocli/testdata/help-deprecated.golden
