+++ client/v2/autocli/testdata/help-echo.golden
