+++ client/v2/autocli/testdata/help-skip.golden
