+++ client/v2/autocli/testdata/help-toplevel.golden
