+++ client/v2/autocli/testdata/help.golden
